+++ src/gdext/utilityfuncs.nim
@@ -2,7 +2,7 @@ import gdext/private/gdinterface
 import gdext/builtinindex
 import gdext/stringtools
 
-proc load(proc_name: string; hash: int): PtrUtilityFunction =
+proc load(proc_name: string; hash: int64): PtrUtilityFunction =
   let name = newStringName proc_name
   interface_Variant_getPtrUtilityFunction(addr name, hash)
